Feat:item Service 추가

+++ src/main/java/com/example/store/service/ItemService.java
@@ -2,6 +2,8 @@ package com.example.store.service;
 
 import com.example.store.domain.Item;
 import com.example.store.dto.ItemDTO;
+import com.example.store.exception.item.InvalidItemException;
+import com.example.store.exception.item.ItemNotFound;
 import com.example.store.repository.ItemRepository;
 import com.example.store.request.ItemCreate;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,7 @@ public class ItemService {
     }
 
     public void addItem(ItemCreate itemCreate) {
+        validateItemCreate(itemCreate);
         Item item = Item.builder()
                 .name(itemCreate.getName())
                 .price(itemCreate.getPrice())
@@ -28,6 +31,9 @@ public class ItemService {
 
     public ItemDTO getItem(Long itemId) {
         Item item = itemRepository.findById(itemId);
+        if (item == null) {
+            throw new ItemNotFound(itemId);
+        }
         return new ItemDTO(item.getItemId(), item.getName(), item.getPrice(), item.getInfo());
     }
 
@@ -39,16 +45,35 @@ public class ItemService {
     }
 
     public void updateItem(Long itemId, ItemCreate itemCreate) {
-        Item item = Item.builder()
-                .itemId(itemId)
-                .name(itemCreate.getName())
-                .price(itemCreate.getPrice())
-                .info(itemCreate.getInfo())
-                .build();
+        Item item = itemRepository.findById(itemId);
+        if (item == null) {
+            throw new ItemNotFound(itemId);
+        }
+        validateItemCreate(itemCreate);
+        item.setName(itemCreate.getName());
+        item.setPrice(itemCreate.getPrice());
+        item.setInfo(itemCreate.getInfo());
         itemRepository.update(item);
     }
 
     public void deleteItem(Long itemId) {
+        Item item = itemRepository.findById(itemId);
+        if (item == null) {
+            throw new ItemNotFound(itemId);
+        }
         itemRepository.delete(itemId);
+    }
+
+    private void validateItemCreate(ItemCreate itemCreate) {
+        if (itemCreate.getName() == null || itemCreate.getName().isEmpty()) {
+            InvalidItemException e = new InvalidItemException("아이템 데이터가 올바르지 않습니다.");
+            e.addValidation("name", "이름이 비어있습니다.");
+            throw e;
+        }
+        if (itemCreate.getPrice() == null || itemCreate.getPrice() <= 0) {
+            InvalidItemException e = new InvalidItemException("아이템 데이터가 올바르지 않습니다.");
+            e.addValidation("price", "가격이 0원보다 커야합니다.");
+            throw e;
+        }
     }
 }
